+++ src/sync/__init__.py
@@ -1,5 +1,0 @@
-"""Synchronization package initialization"""
-
-from .sync_data import LyricWord, LyricLine, LyricsData
-
-__all__ = ['LyricWord', 'LyricLine', 'LyricsData']